refactor to move to loader + first working calls to versioned extension from loader
