+++ src/timescale_bgw.h
@@ -1,0 +1,3 @@
+
+
+extern void timescale_bgw_scheduler_main(Oid db_id);
